Refactor Inventory List: Split components into separate files
- Replaced InventoryList component with InventoryListTab for better organization.
- Moved InventoryList logic into InventoryListTab.tsx.
- Created InventoryCards, InventoryFilters, and InventoryTable components for modularity.
- Updated imports and exports accordingly.
- Removed the old InventoryList component file.
