+++ src/pages/inventory/tabs/List/components/index.ts
@@ -1,0 +1,3 @@
+export { InventoryFilters } from "./InventoryFilters";
+export { InventoryTable } from "./InventoryTable";
+export { InventoryCards } from "./InventoryCards";
